Ignore read-only mappings of ELF files at non-fixed addresses.
Such a mapping is likely part of some transient operation by the dynamic
linker, such as parsing the ELF header while loading an object.  If the
address of that mapping is recycled it may trigger disposal of the
corresponding debug info since it overlaps with a new mapping.  The
result of this is that valgrind is unable to hook malloc() calls in the
client, so memcheck reports zero heap usage.

Fix the problem by ignoring read-only non-fixed ELF file mappings.  This
might not be the best way to address the problem, but it seems to work.
Older versions of valgrind ignored read-only mappings completely.